Forgot to add the changes involving moving the API file.
