+++ src/components/APIButton.tsx
@@ -7,35 +7,32 @@ import {openai} from "./Home"
 export function Counter(): JSX.Element {
   function inputAPI() {
     
-    //const openai = new OpenAI();
-
     async function computeAPI() {
       const completion = await openai.chat.completions.create({
         messages: [{ role: "system", content: "Come up with a rap." }],
         model: "gpt-3.5-turbo",
       });
       
-      setValue(JSON.stringify(completion.choices[0]));
+      setValue(JSON.stringify(completion.choices[0]["message"]["content"]).replace(/\\n/g, "\n"));
       console.log(completion.choices[0]);
     
     }
     computeAPI();
   }
+  const [num, setNum] = useState<number>(0);
   const [value, setValue] = useState<string>("");
   return (
     <div>
       <span>
-          <Button onClick={() => setValue(1 + value)}>Add Two</Button>
-          to {value}.
+          <Button onClick={() => setNum(1 + num)}>Add Two</Button>
+          to {num}.
       </span>
       <span>
           <Button onClick={() => inputAPI()}> Here is the API stuff </Button>
       </span>
       <span>
-        <div>
-            The current value is:{value}
-          </div>
-        </span>
+          The current value is:{value}
+      </span>
     </div>
   );
 }
